+++ src/grepros/main.py
@@ -271,9 +271,9 @@ Export all bag messages to SQLite and Postgres, print only export progress:
                   "  template=/my/path.tpl - custom template to use for HTML output\n"
                   "  commit-interval=NUM - transaction size for Postgres output\n"
                   "                        (default 100, 0 is autocommit)\n"
-                  "  nesting=lists|all - create tables for nested message types\n"
+                  "  nesting=array|all - create tables for nested message types\n"
                   "                      in Postgres/SQLite output,\n"
-                  '                      only for arrays if "lists" else for any nested types\n'
+                  '                      only for arrays if "array" else for any nested types\n'
                   "                      (array fields in parent will be populated with foreign keys\n"
                   "                       instead of formatted nested values)"),