+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog for package grepros
 
 0.3.2 (2021-12-08)
 -------------------
-* add --write-option subtypes=lists|all, for Postgres/SQLite output
+* add --write-option subtypes=array|all, for Postgres/SQLite output
 * rename --write-option commit_interval to commit-interval
 * raise Postgres default commit-interval from 100 to 1000